% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-bugs.R
\name{bugs.worker}
\alias{bugs.worker}
\title{Worker function, runs on each CPU}
\usage{
bugs.worker(chain, data, model.str, inits, params, n.iter, n.burnin,
  OpenBUGS.exe, bugs.debug = FALSE)
}
\arguments{
\item{chain}{number the chain is (1, 2, 3, ...)}

\item{data}{the R2OpenBUGS::bugs data}

\item{model.str}{the OpenBugs model string}

\item{inits}{the inits list from the main function}

\item{params}{the params to record}

\item{n.iter}{number of iterations}

\item{n.burnin}{number of burnin iterations}

\item{OpenBUGS.exe}{the openbugs executable location}

\item{bugs.debug}{run R2OpenBUGS with debug flag}
}
\value{
coda mcmc data
}
\description{
Worker function, runs on each CPU
}

