% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-bugs.R
\name{build.cluster.list}
\alias{build.cluster.list}
\title{Builds a list of available CPUs for sfInit.}
\usage{
build.cluster.list(host.list, cpu.counts)
}
\arguments{
\item{host.list}{list of ip addresses, or a single ip address}

\item{cpu.counts}{list of corresponding available cpus, or cpu count for all machines, or cpu count for single machine}
}
\value{
list of available CPUs

build.cluster.list('localhost', 3)   # c('localhost', 'localhost', 'localhost')
build.cluster.list(c('machine1', 'machine2', 'machine3'), 1)  # c('machine1', 'machine2', 'machine3')
build.cluster.list(c('machine1', 'machine2'), c(2, 1))   # c('machine1', 'machine1', 'machine2')
}
\description{
Builds a list of available CPUs for sfInit.
}

