% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-bugs.R
\name{parallel.bugs}
\alias{parallel.bugs}
\title{Main function of parallelBugs. Pass parameters that would be passed to R2OpenBUGS::bugs, as well as
cpu/host information, to run the model with multiple chains in parallel.}
\usage{
\method{parallel}{bugs}(host.list, cpu.counts, model.str, bugs.data, n.chains,
  init.list, parameters.to.save, OpenBUGS.exe, n.iter = 5000,
  n.burnin = 1000, bugs.debug = FALSE)
}
\arguments{
\item{host.list}{A single ip as a string ('localhost', 'server.org') or a vector of string ips c('machine1', 'machine2', 'machine3')}

\item{cpu.counts}{A single integer if each machine has the same number of cores to utilize, or a vector corresponding to the number of cores to utilize per machine}

\item{model.str}{the OpenBUGS model as a string}

\item{bugs.data}{data to pass to R2OpenBUGS::bugs - it should be a list}

\item{n.chains}{The number of chains to run - it should equal the sum of cpu.counts}

\item{init.list}{list of inits for each chain, structured as list(chain1=list(...), chain2=(list(...)), ...)}

\item{parameters.to.save}{params to pass to R2OpenBUGS - should be a list}

\item{OpenBUGS.exe}{OpenBUGS executable location, to be passed to OpenBUGS.pgm}

\item{n.iter}{number of iterations}

\item{n.burnin}{number of burnin iterations}

\item{bugs.debug}{if R2OpenBUGS will run bugs() call with debug true}
}
\value{
coda mcmc.list object
}
\description{
Main function of parallelBugs. Pass parameters that would be passed to R2OpenBUGS::bugs, as well as
cpu/host information, to run the model with multiple chains in parallel.
}
\examples{
model.str = "model
{
 tau ~ dgamma(0.0001, 0.0001)
 mu ~ dflat()
 sigma <- 1 / sqrt(tau)

 for(i in 1:N)
 {
   x.data[i]~dnorm(mu, tau)
 }
}"
bugs.data = list(x.data=rnorm(1000, 12, 5), N=1000)

# Initial values of parameters. Note that you can use
# R random number functions here
init.list = list(
 chain1=list(mu=runif(1, -50, 50), tau=1),
 chain2=list(mu=-50, tau=10),
 chain3=list(mu=50, tau=.1)
)

# Parameters to collect statistics on
params <- c("mu", "sigma")

OpenBUGS.exe="C:/Program Files (x86)/OpenBUGS/OpenBUGS323/OpenBUGS.exe"

\dontrun{
parallel.bugs('localhost', 3, model.str,
             bugs.data, 3, init.list, params,
             OpenBUGS.exe, n.iter=1000)
}
}

